Rename to avoid PyLance warning
